+++ Flightwise/app/src/main/java/by/bsu/flightwise/ui/activities/AboutActivity.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.tooling.preview.Preview
 import by.bsu.flightwise.R
+import by.bsu.flightwise.ui.fragments.HeaderFragment
 import by.bsu.flightwise.ui.theme.FlightwiseTheme
 
 class AboutActivity : ComponentActivity() {
@@ -46,86 +47,7 @@ fun AboutScreen() {
 
     Box(modifier = Modifier.fillMaxSize()) {
 
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(MaterialTheme.colorScheme.primary)
-                .padding(16.dp)
-                .align(Alignment.TopCenter)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "Flightwise",
-                    color = MaterialTheme.colorScheme.onPrimary,
-                    style = MaterialTheme.typography.headlineMedium,
-                    modifier = Modifier.weight(1f)
-                )
-
-                Box {
-                    IconButton(onClick = { expanded = !expanded }) {
-                        Icon(
-                            imageVector = Icons.Default.Menu,
-                            contentDescription = "Menu",
-                            tint = MaterialTheme.colorScheme.onPrimary
-                        )
-                    }
-
-                    DropdownMenu(
-                        expanded = expanded,
-                        onDismissRequest = { expanded = false },
-                        modifier = Modifier
-                            .background(MaterialTheme.colorScheme.secondary)
-                            .width(160.dp)
-                    ) {
-                        val themeColors = MenuDefaults.itemColors(
-                            textColor = MaterialTheme.colorScheme.onSecondary,
-                            leadingIconColor = MaterialTheme.colorScheme.secondary,
-                            trailingIconColor = MaterialTheme.colorScheme.tertiary,
-                            disabledTextColor = MaterialTheme.colorScheme.onTertiary,
-                            disabledLeadingIconColor = MaterialTheme.colorScheme.onTertiary,
-                            disabledTrailingIconColor = MaterialTheme.colorScheme.onTertiary
-                        )
-
-                        DropdownMenuItem(
-                            onClick = {
-                                //context.startActivity(Intent(context, ProfileActivity::class.java))
-                                expanded = false
-                            },
-                            text = { Text(stringResource(id = R.string.menu_profile), style = MaterialTheme.typography.labelMedium) },
-                            enabled = true,
-                            modifier = Modifier.background(MaterialTheme.colorScheme.secondary),
-                            colors = themeColors,
-                            contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp)
-                        )
-                        DropdownMenuItem(
-                            onClick = {
-                                //context.startActivity(Intent(context, TicketsActivity::class.java))
-                                expanded = false
-                            },
-                            text = { Text(stringResource(id = R.string.menu_tickets), style = MaterialTheme.typography.labelMedium) },
-                            enabled = true,
-                            modifier = Modifier.background(MaterialTheme.colorScheme.secondary),
-                            colors = themeColors,
-                            contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp)
-                        )
-                        DropdownMenuItem(
-                            onClick = {
-                                //context.startActivity(Intent(context, CountriesActivity::class.java))
-                                expanded = false
-                            },
-                            text = { Text(stringResource(id = R.string.menu_countries), style = MaterialTheme.typography.labelMedium) },
-                            enabled = true,
-                            modifier = Modifier.background(MaterialTheme.colorScheme.secondary),
-                            colors = themeColors,
-                            contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp)
-                        )
-                    }
-                }
-            }
-        }
+        HeaderFragment()
 
         Column(
             modifier = Modifier
